+++ miles_plugins/mbridge/qwen3_next.py
@@ -1,58 +1,7 @@
-from argparse import Namespace
-
 import torch
 
 from mbridge.core import register_model
 from mbridge.models import Qwen2MoEBridge
-
-
-def needs_gdn_weight_fix(name: str) -> bool:
-    """Check if a parameter name requires GDN weight gather fix."""
-    if "self_attention.in_proj.weight" in name and "layer_norm" not in name:
-        return True
-    if "self_attention.conv1d.weight" in name:
-        return True
-    if "self_attention.conv1d.bias" in name:
-        return True
-    return False
-
-
-def fix_gdn_weight_gather(args: Namespace, name: str, param: torch.Tensor) -> torch.Tensor:
-    """Fix GDN (GatedDeltaNet) per-component TP gathering.
-
-    MCore's sharded_state_dict loads GDN in_proj and conv1d weights with per-component
-    TP sharding: each rank holds [Q_local, K_local, V_local, ...].  But partition_stride=1
-    causes all_gather_param to do a simple cat, producing
-        [Q_r0, K_r0, V_r0, ..., Q_r1, K_r1, V_r1, ...]
-    instead of the correct layout
-        [Q_all, K_all, V_all, ...].
-    This function rearranges the gathered tensor to the correct component-contiguous layout.
-    """
-    from megatron.core import mpu
-
-    tp_size = mpu.get_tensor_model_parallel_world_size()
-    if tp_size <= 1:
-        return param
-
-    qk_local = args.linear_num_key_heads * args.linear_key_head_dim // tp_size
-    v_local = args.linear_num_value_heads * args.linear_value_head_dim // tp_size
-    nv_local = args.linear_num_value_heads // tp_size
-
-    if "self_attention.in_proj.weight" in name and "layer_norm" not in name:
-        sections = [qk_local, qk_local, v_local, v_local, nv_local, nv_local]
-    elif "self_attention.conv1d.weight" in name:
-        sections = [qk_local, qk_local, v_local]
-    elif "self_attention.conv1d.bias" in name:
-        sections = [qk_local, qk_local, v_local]
-    else:
-        return param
-
-    chunks = param.chunk(tp_size, dim=0)
-    per_rank_comps = [c.split(sections, dim=0) for c in chunks]
-    return torch.cat(
-        [torch.cat([per_rank_comps[r][c] for r in range(tp_size)], dim=0) for c in range(len(sections))],
-        dim=0,
-    )
 
 
 def convert_gated_attn_qgkv_mcore_to_hf(mcore_weights, *, num_heads, num_kv_heads, head_dim, hidden_size):
@@ -176,13 +125,6 @@ class Qwen3NextBridge(Qwen2MoEBridge):
         if "self_attention.in_proj." in mcore_weights_name and "layer_norm" not in mcore_weights_name:
             return self._convert_gdn_in_proj_hf_to_mcore(hf_weights)
 
-        if "self_attention.out_norm.weight" in mcore_weights_name:
-            assert len(hf_weights) == 1
-            # Qwen3-Next mixed norm: linear_attn.norm is direct-scale in HF, but
-            # Megatron's --apply-layernorm-1p computes (1+w). Subtract 1 to compensate.
-            # FP32 arithmetic then cast back to bf16 (1-2 ULP round-trip error is accepted).
-            return (hf_weights[0].float() - 1.0).bfloat16()
-
         # Gated attention: MCore still names the attribute "linear_qkv" even with attention_output_gate=True,
         # but the weight contains [Q, G, K, V] per group instead of [Q, K, V].
         if "self_attention.linear_qkv." in mcore_weights_name and "layer_norm" not in mcore_weights_name:
@@ -197,12 +139,6 @@ class Qwen3NextBridge(Qwen2MoEBridge):
             hf_names = self._weight_name_mapping_mcore_to_hf(mcore_weights_name)
             hf_weights = self._convert_gdn_in_proj_mcore_to_hf(mcore_weights)
             return hf_names, hf_weights
-
-        if "self_attention.out_norm.weight" in mcore_weights_name:
-            hf_names = self._weight_name_mapping_mcore_to_hf(mcore_weights_name)
-            # Reverse the -1 applied during hf→mcore: restore to HF direct-scale format.
-            # Use FP32 arithmetic to preserve precision (out_norm is stored in FP32).
-            return hf_names, [(mcore_weights.float() + 1.0).bfloat16()]
 
         if "self_attention.linear_qkv." in mcore_weights_name and "layer_norm" not in mcore_weights_name:
             hf_names = self._weight_name_mapping_mcore_to_hf(mcore_weights_name)
